requirements.txt
python==3.7.10
jupyter==1.0.0
tensorflow==2.4.1
scikit-image==0.16.2
opencv-contrib-python==4.5.1.48
numpy~=1.18.5
matplotlib~=3.3.4
pillow~=8.1.2